добавляет карту на страницу offer
