+++ src/components/app/app.tsx
@@ -1,7 +1,7 @@
 import MainPage from '../../pages/main-page';
 import FavoritePage from '../../pages/favorites-page';
 import LoginPage from '../../pages/login-page';
-import OfferPage from '../../pages/offer-page';
+import OfferPage from '../../pages/offer-page/offer-page';
 import NotFoundPage from '../../pages/not-found-page';
 import PrivateRoute from '../private-route';
 import Layout from '../layout/layout';
